5/20

+++ src/controllers/video.controller.ts
@@ -1,4 +1,5 @@
 import { Request, Response, NextFunction } from "express";
+import User from "../models/User";
 import Video from "../models/Video";
 import formatHashtags from "../utils/formatHashtags";
 
@@ -10,7 +11,7 @@ export const home = async (req: Request, res: Response) => {
 export const watch = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if (!video) {
       return res.render("404", { pageTitle: "Video not found" });
     }
@@ -20,19 +21,27 @@ export const watch = async (req: Request, res: Response) => {
 
 export const getEdit = async (req: Request, res: Response) => {
   const { id } = req.params;
+  const { user } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== String(user?._id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit", { pageTitle: "Editing", video });
 };
 
 export const postEdit = async (req: Request, res: Response) => {
   const { id } = req.params;
+  const { user } = req.session;
   const { title, description, hashtags } = req.body;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== String(user?._id)) {
+    return res.status(403).redirect("/");
   }
   video.title = title;
   video.description = description;
@@ -47,9 +56,15 @@ export const getUpload = (req: Request, res: Response) => {
 
 export const postUpload = async (req: Request, res: Response) => {
   try {
+    const { path: fileUrl } = req.file;
     const { title, description, hashtags } = req.body;
-    await Video.create({
+    const {
+      session: { user },
+    } = req;
+    const newVideo = await Video.create({
       title,
+      fileUrl,
+      owner: user?._id,
       description,
       createdAt: Date.now(),
       hashtags: formatHashtags(hashtags),
@@ -58,6 +73,9 @@ export const postUpload = async (req: Request, res: Response) => {
         rating: 0,
       },
     });
+    const saveUser = await User.findById(user?._id);
+    saveUser?.videos.push(newVideo._id);
+    await saveUser?.save();
     return res.redirect("/");
   } catch (error) {
     console.log(error);
@@ -76,12 +94,17 @@ export const search = async (req: Request, res: Response) => {
   return res.render("search", { pageTitle: "Search", videos });
 };
 
-export const upload = (req: Request, res: Response) => res.send("upload");
-
 export const deleteVideo = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+    const { user } = req.session;
     const video = await Video.findById(id);
+    if (!video) {
+      return res.status(404).render("404", { pageTitle: "Video not found" });
+    }
+    if (String(video.owner) !== String(user?._id)) {
+      return res.status(403).redirect("/");
+    }
     await Video.findOneAndDelete(video?.id);
     return res.redirect("/");
   } catch (error) {
